Move textfield listener to new package
