+++ core/src/test/java/se2/ticktackbumm/core/gamelogic/TextfieldInputListenerTest.java
@@ -1,4 +1,4 @@
-package se2.ticktackbumm.core.screens;
+package se2.ticktackbumm.core.gamelogic;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class TextfieldInputListenerTest {
-
     TextfieldInputListener textfieldInputListener;
 
     @BeforeEach
